# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Grupo Beija-flor Internal Social Network API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      parameters:
        - name: tag
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/avatar:
    post:
      operationId: updateAvatar
      tags: [users]
      summary: Update user avatar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAvatarRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/allowed-emails:
    get:
      operationId: listAllowedEmails
      tags: [users]
      summary: List allowed emails
      responses:
        "200":
          description: List of allowed emails
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AllowedEmail"
    post:
      operationId: addAllowedEmail
      tags: [users]
      summary: Add allowed email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddAllowedEmailRequest"
      responses:
        "200":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllowedEmail"
  /admin/allowed-emails/{id}:
    delete:
      operationId: removeAllowedEmail
      tags: [users]
      summary: Remove allowed email
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: List tickets
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, in_progress, closed]
      responses:
        "200":
          description: Tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketRequest"
      responses:
        "200":
          description: Created ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
    patch:
      operationId: updateTicketStatus
      tags: [tickets]
      summary: Update ticket status (assigned team)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTicketRequest"
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}/messages:
    get:
      operationId: listTicketMessages
      tags: [tickets]
      summary: List ticket messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TicketMessage"
    post:
      operationId: addTicketMessage
      tags: [tickets]
      summary: Add message to ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketMessageRequest"
      responses:
        "200":
          description: Message added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketMessage"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    UpdateAvatarRequest:
      type: object
      properties:
        avatarUrl:
          type: string
      required:
        - avatarUrl
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: string
          nullable: true
        tag:
          type: string
          enum: [marketing, adm, socio, posto, churrascaria, gerente]
          nullable: true
        role:
          type: string
          enum: [user, moderator, admin, master_admin]
        birthDate:
          type: string
          nullable: true
        admissionDate:
          type: string
          nullable: true
        onboardingCompleted:
          type: boolean
        acceptedTerms:
          type: boolean
        readDocuments:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - onboardingCompleted
        - acceptedTerms
        - readDocuments
        - createdAt
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        tag:
          type: string
          enum: [marketing, adm, socio, posto, churrascaria, gerente]
        role:
          type: string
          enum: [user, moderator, admin, master_admin]
        birthDate:
          type: string
          nullable: true
        admissionDate:
          type: string
          nullable: true
    AllowedEmail:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        addedAt:
          type: string
      required:
        - id
        - email
        - addedAt
    AddAllowedEmailRequest:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        tag:
          type: string
          enum: [marketing, adm, socio, posto, churrascaria, gerente]
        role:
          type: string
          enum: [user, moderator, admin, master_admin]
        temporaryPassword:
          type: string
      required:
        - email
        - name
        - temporaryPassword
...
    UserBasic:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        avatarUrl:
          type: string
          nullable: true
        tag:
          type: string
          nullable: true
        role:
          type: string
      required:
        - id
        - name
        - role
...
    Ticket:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in_progress, closed]
        category:
          type: string
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/UserBasic"
        messageCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - description
        - status
        - category
        - authorId
        - author
        - messageCount
        - createdAt
        - updatedAt
    CreateTicketRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
      required:
        - title
        - description
        - category
    UpdateTicketRequest:
      type: object
      properties:
        status:
          type: string
          enum: [open, in_progress, closed]
    TicketMessage:
      type: object
      properties:
        id:
          type: integer
        ticketId:
          type: integer
        content:
          type: string
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/UserBasic"
        createdAt:
          type: string
      required:
        - id
        - ticketId
        - content
        - authorId
        - author
        - createdAt
    CreateTicketMessageRequest:
      type: object
      properties:
        content:
          type: string
      required:
        - content
...